pyproject.toml
[project]
name = "credit-ocr-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.30.1",
    "pip>=25.2",
    "requests>=2.32.4",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "psutil>=5.9.0",
    "matplotlib>=3.10.5",
    "pdf2image>=1.17.0",
    "pymupdf>=1.26.3",
    "aiohttp>=3.12.15",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "setuptools>=80.9.0",
    "azure-storage-blob>=12.19.0,<12.22.0",
    "testcontainers[postgres]>=4.9.0",
    "celery>=5.3.0",
    "easyocr>=1.7.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
